+++ lib/technicaldetails.dart
@@ -4,8 +4,10 @@ import 'professionaldetails.dart';
 import 'educationdetails.dart';
 import 'edittechnical.dart';
 import 'main.dart';
+import 'package:connectivity/connectivity.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'auth.dart';
+import 'welcomescreen.dart';
 import 'updatetechnical.dart';
 import 'package:flutter/scheduler.dart' show timeDilation;
 
@@ -86,16 +88,43 @@ class _TechnicalDetailsState extends State<TechnicalDetails> {
     progress(false);
     return Dismissible(
         key: Key(document.documentID),
-        onDismissed: (DismissDirection direction) {
-          direction = DismissDirection.endToStart;
-
-          Firestore.instance
-              .collection('users')
-              .document('technical')
-              .collection(userid)
-              .document(document.documentID)
-              .delete();
+        confirmDismiss: (direction) async {
+          final bool res = await showDialog(
+              context: context,
+              builder: (BuildContext context) {
+                return AlertDialog(
+                  title: const Text("Confirm"),
+                  content: const Text("Do you want to delete this item?"),
+                  actions: <Widget>[
+                    FlatButton(
+                        onPressed: () {
+                          Navigator.of(context).pop(true);
+                          Firestore.instance
+                              .collection('users')
+                              .document('technical')
+                              .collection(userid)
+                              .document(document.documentID)
+                              .delete();
+                        },
+                        child: const Text("DELETE")),
+                    FlatButton(
+                      onPressed: () => Navigator.of(context).pop(false),
+                      child: const Text("CANCEL"),
+                    )
+                  ],
+                );
+              });
         },
+        background: Container(
+          alignment: Alignment(0.8, 0),
+          color: Color(0xffe8eaed),
+          child: Icon(
+            Icons.delete,
+            color: Colors.grey,
+            size: 35.0,
+          ),
+        ),
+        direction: DismissDirection.endToStart,
         child: Card(
           child: Padding(
             padding: const EdgeInsets.only(top: 8.0, bottom: 8.0),
@@ -126,7 +155,8 @@ class _TechnicalDetailsState extends State<TechnicalDetails> {
                     ),
                     Row(
                       children: <Widget>[
-                        Text(document['startdate'] + ' - ' + document['enddate'],
+                        Text(
+                            document['startdate'] + ' - ' + document['enddate'],
                             style: TextStyle(
                                 fontSize: 16.0,
                                 color: Colors.blueGrey,
@@ -152,6 +182,7 @@ class _TechnicalDetailsState extends State<TechnicalDetails> {
               trailing: IconButton(
                   icon: Icon(Icons.edit),
                   onPressed: () {
+                    checkConnectivity();
                     Navigator.push(
                         context,
                         MaterialPageRoute(
@@ -170,6 +201,44 @@ class _TechnicalDetailsState extends State<TechnicalDetails> {
         ));
   }
 
+
+  checkConnectivity() async {
+    var result = await Connectivity().checkConnectivity();
+    if (result == ConnectivityResult.none) {
+      showDialog(
+          barrierDismissible: false,
+          context: context,
+          builder: (BuildContext context) {
+            return WillPopScope(
+              onWillPop: ()async=> false,
+              child: AlertDialog(
+                title: Text('Oops! Internet lost'),
+                content: Text(
+                    'Sorry, Please ckeck your internet connection and then try again'),
+                actions: <Widget>[
+                  FlatButton(
+                    child: Text('OK'),
+                    onPressed: () {
+                      checkConnectivity();
+                      Navigator.pop(context);
+                     // Navigator.pushReplacement(context, MaterialPageRoute(builder: (context)=>WelcomeScreen()));
+
+                    },
+                  )
+                ],
+
+
+              ),
+            );
+          });
+    } else if (result == ConnectivityResult.mobile) {
+
+    } else if (result == ConnectivityResult.wifi) {
+
+    }
+  }
+
+
   Widget progress(bool visibility) {
     return Visibility(
         child: CircularProgressIndicator(
@@ -182,6 +251,8 @@ class _TechnicalDetailsState extends State<TechnicalDetails> {
   void initState() {
     // TODO: implement initState
     super.initState();
+    timeDilation = 2.0;
+    checkConnectivity();
     widget.authService.getCurrentuser().then((userid) {
       setState(() {
         this.userid = userid;
@@ -207,6 +278,7 @@ class _TechnicalDetailsState extends State<TechnicalDetails> {
                   color: Colors.white,
                 ),
                 onPressed: () {
+                  checkConnectivity();
                   Navigator.push(
                       context,
                       MaterialPageRoute(
@@ -220,10 +292,11 @@ class _TechnicalDetailsState extends State<TechnicalDetails> {
         ),
         body: Column(
           children: <Widget>[
-          SizedBox(height: 10,),
+            SizedBox(
+              height: 10,
+            ),
             Expanded(
               child: Container(
-
                 width: MediaQuery.of(context).size.width,
                 height: MediaQuery.of(context).size.height,
                 child: StreamBuilder(
